Moved test_vdi_restructuring.py into test_data.py #297
